WIP: Check if distro is running
*	modified:   CHANGELOG.md
*	modified:   src/backup-WSL.ps1

+++ src/backup-WSL.ps1
@@ -25,7 +25,7 @@ PARAM (
 #--------------------------------------------------------------------------------------------
 # START SET
 $debug = 1
-$version = "0.2.1"
+$version = "0.2.2"
 $app = "backup-WSL.ps1"
 $info = "backup WSL"
 $ld = "c:\tmp\log\"
@@ -55,18 +55,26 @@ function waitforenter {
 function check($distroparam) {
   $distroparam
   $wslinfo = wsl --list --running
+  # $singlelinewslrunning = $wslinfo -replace "`r?`n(?!`r?`n)", ''
+  $arrwslinfo = $wslinfo.Split("`r`n")
+  # check if multiple distros are spinning
+Write-Host "Total Elements in array-->" $arrwslinfo.Count
+Write-Host "First element in array-->" $arrwslinfo[0]
+Write-Host "Second element in array-->" $arrwslinfo[2]
   if ($wslinfo -eq 'Er zijn geen actieve distributies.') { 
     "continue" 
       " Continue script " 
   }
   else {
     "--START WSL INFO --"
-    $wslinfo
+    $arrwslinfo[2]
     "-- END  WSL INFO --"
       " IF $distroparam IS Running "
+      if ($arrwslinfo[2] -contains $distroparam) {
       " HALT script "
       " Send poweroff to running distro "
       "sudo systemctl poweroff"
+      } 
      waitforenter
   }
 }
